removes sent counter as value from spammed transactions

+++ plugins/webapi/spammer/plugin.go
@@ -61,5 +61,5 @@ type Response struct {
 
 type Request struct {
 	Cmd string `json:"cmd"`
-	Tps uint   `json:"tps"`
+	Tps uint64 `json:"tps"`
 }
